- Saving progress on Sound 4 again.  It's still wrong. refs #39
- WaveOutSound's playback thread wasn't checking for application exit in one spot, causing a crash on exit.
refs #41

+++ Source/sound/sound.h
@@ -184,7 +184,7 @@ private:
 	double m_sound4StartTimeSeconds;
 	double m_sound4LengthSeconds;
 
-	u8 m_sound4Sample;
+	u32 m_sound4Sample;
 
 	bool m_envelope4Enabled;
 	double m_lastEnvelope4UpdateTimeSeconds;
